Alternating Optimization

This function performs alternating optimization on the function f.

ao(f, npar, groups, sequence, initial, minimize = TRUE, progress = FALSE, ...)

Arguments

f
A function of \(n\) variables to be optimized.

npar
An integer, the number \(n\) of variables of f.

groups
A list of vectors of parameter indices \(1,...,n\) of f. This determines the grouping of parameters. Indices can be present in multiple groups. If one group equals \(1,...,n\), full estimation is carried out.

sequence
A vector of indices of the list groups. This determines the sequence in which parameter groups get optimized.

initial
A vector of length \(n\) of initial parameter values. If not supplied, they are randomly drawn.

minimize
A boolean, determining whether to minimize (minimize = TRUE) or to maximize (minimize = FALSE) the function f.

progress
A boolean, determining whether progress should be printed.

...
Arguments that get passed on to nlm.

Value

A list containing the following components:

optimum

The optimal value of f.

estimate

The parameter vector at which the optimum of f is obtained.

time

The total optimization time in seconds.

nlm_outputs

A list of nlm outputs in each sequence.

minimize

The input minimize.

Examples

f = function(x) 3*x[1]^2 + 2*x[1]*x[2] + x[2]^2 - 5*x[1] + 2
npar = 2
sequence = rep(c(1,2),10)
groups = list(1,2)
ao(f = f, npar = npar, sequence = sequence, groups = groups)
#> $optimum
#> [1] -1.125
#> 
#> $estimate
#>       x_1       x_2 
#>  1.249975 -1.249975 
#> 
#> $time
#> Time difference of 0.002025843 secs
#> 
#> $nlm_outputs
#> $nlm_outputs[[1]]
#> $nlm_outputs[[1]]$minimum
#> [1] 0.385653
#> 
#> $nlm_outputs[[1]]$estimate
#> [1] 0.7482272
#> 
#> $nlm_outputs[[1]]$gradient
#> [1] 1.998401e-07
#> 
#> $nlm_outputs[[1]]$code
#> [1] 1
#> 
#> $nlm_outputs[[1]]$iterations
#> [1] 1
#> 
#> 
#> $nlm_outputs[[2]]
#> $nlm_outputs[[2]]$minimum
#> [1] -0.6214481
#> 
#> $nlm_outputs[[2]]$estimate
#> [1] -0.7482272
#> 
#> $nlm_outputs[[2]]$gradient
#> [1] 0
#> 
#> $nlm_outputs[[2]]$code
#> [1] 1
#> 
#> $nlm_outputs[[2]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[3]]
#> $nlm_outputs[[3]]$minimum
#> [1] -0.9571494
#> 
#> $nlm_outputs[[3]]$estimate
#> [1] 1.082742
#> 
#> $nlm_outputs[[3]]$gradient
#> [1] -6.152283e-12
#> 
#> $nlm_outputs[[3]]$code
#> [1] 1
#> 
#> $nlm_outputs[[3]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[4]]
#> $nlm_outputs[[4]]$minimum
#> [1] -1.06905
#> 
#> $nlm_outputs[[4]]$estimate
#> [1] -1.082742
#> 
#> $nlm_outputs[[4]]$gradient
#> [1] 2.050761e-12
#> 
#> $nlm_outputs[[4]]$code
#> [1] 1
#> 
#> $nlm_outputs[[4]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[5]]
#> $nlm_outputs[[5]]$minimum
#> [1] -1.10635
#> 
#> $nlm_outputs[[5]]$estimate
#> [1] 1.194247
#> 
#> $nlm_outputs[[5]]$gradient
#> [1] 0
#> 
#> $nlm_outputs[[5]]$code
#> [1] 1
#> 
#> $nlm_outputs[[5]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[6]]
#> $nlm_outputs[[6]]$minimum
#> [1] -1.118783
#> 
#> $nlm_outputs[[6]]$estimate
#> [1] -1.194247
#> 
#> $nlm_outputs[[6]]$gradient
#> [1] 3.718569e-12
#> 
#> $nlm_outputs[[6]]$code
#> [1] 1
#> 
#> $nlm_outputs[[6]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[7]]
#> $nlm_outputs[[7]]$minimum
#> [1] -1.122928
#> 
#> $nlm_outputs[[7]]$estimate
#> [1] 1.231416
#> 
#> $nlm_outputs[[7]]$gradient
#> [1] -3.60633e-12
#> 
#> $nlm_outputs[[7]]$code
#> [1] 1
#> 
#> $nlm_outputs[[7]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[8]]
#> $nlm_outputs[[8]]$minimum
#> [1] -1.124309
#> 
#> $nlm_outputs[[8]]$estimate
#> [1] -1.231416
#> 
#> $nlm_outputs[[8]]$gradient
#> [1] -1.803165e-11
#> 
#> $nlm_outputs[[8]]$code
#> [1] 3
#> 
#> $nlm_outputs[[8]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[9]]
#> $nlm_outputs[[9]]$minimum
#> [1] -1.12477
#> 
#> $nlm_outputs[[9]]$estimate
#> [1] 1.243805
#> 
#> $nlm_outputs[[9]]$gradient
#> [1] 0
#> 
#> $nlm_outputs[[9]]$code
#> [1] 1
#> 
#> $nlm_outputs[[9]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[10]]
#> $nlm_outputs[[10]]$minimum
#> [1] -1.124923
#> 
#> $nlm_outputs[[10]]$estimate
#> [1] -1.243805
#> 
#> $nlm_outputs[[10]]$gradient
#> [1] -1.785204e-12
#> 
#> $nlm_outputs[[10]]$code
#> [1] 1
#> 
#> $nlm_outputs[[10]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[11]]
#> $nlm_outputs[[11]]$minimum
#> [1] -1.124974
#> 
#> $nlm_outputs[[11]]$estimate
#> [1] 1.247935
#> 
#> $nlm_outputs[[11]]$gradient
#> [1] -3.558592e-12
#> 
#> $nlm_outputs[[11]]$code
#> [1] 1
#> 
#> $nlm_outputs[[11]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[12]]
#> $nlm_outputs[[12]]$minimum
#> [1] -1.124991
#> 
#> $nlm_outputs[[12]]$estimate
#> [1] -1.247935
#> 
#> $nlm_outputs[[12]]$gradient
#> [1] -1.779296e-12
#> 
#> $nlm_outputs[[12]]$code
#> [1] 1
#> 
#> $nlm_outputs[[12]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[13]]
#> $nlm_outputs[[13]]$minimum
#> [1] -1.124997
#> 
#> $nlm_outputs[[13]]$estimate
#> [1] 1.249312
#> 
#> $nlm_outputs[[13]]$gradient
#> [1] -1.777336e-12
#> 
#> $nlm_outputs[[13]]$code
#> [1] 1
#> 
#> $nlm_outputs[[13]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[14]]
#> $nlm_outputs[[14]]$minimum
#> [1] -1.124999
#> 
#> $nlm_outputs[[14]]$estimate
#> [1] -1.249312
#> 
#> $nlm_outputs[[14]]$gradient
#> [1] 0
#> 
#> $nlm_outputs[[14]]$code
#> [1] 1
#> 
#> $nlm_outputs[[14]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[15]]
#> $nlm_outputs[[15]]$minimum
#> [1] -1.125
#> 
#> $nlm_outputs[[15]]$estimate
#> [1] 1.249771
#> 
#> $nlm_outputs[[15]]$gradient
#> [1] 0
#> 
#> $nlm_outputs[[15]]$code
#> [1] 1
#> 
#> $nlm_outputs[[15]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[16]]
#> $nlm_outputs[[16]]$minimum
#> [1] -1.125
#> 
#> $nlm_outputs[[16]]$estimate
#> [1] -1.249771
#> 
#> $nlm_outputs[[16]]$gradient
#> [1] 1.776683e-12
#> 
#> $nlm_outputs[[16]]$code
#> [1] 1
#> 
#> $nlm_outputs[[16]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[17]]
#> $nlm_outputs[[17]]$minimum
#> [1] -1.125
#> 
#> $nlm_outputs[[17]]$estimate
#> [1] 1.249924
#> 
#> $nlm_outputs[[17]]$gradient
#> [1] -7.105862e-12
#> 
#> $nlm_outputs[[17]]$code
#> [1] 1
#> 
#> $nlm_outputs[[17]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[18]]
#> $nlm_outputs[[18]]$minimum
#> [1] -1.125
#> 
#> $nlm_outputs[[18]]$estimate
#> [1] -1.249924
#> 
#> $nlm_outputs[[18]]$gradient
#> [1] -3.552931e-12
#> 
#> $nlm_outputs[[18]]$code
#> [1] 1
#> 
#> $nlm_outputs[[18]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[19]]
#> $nlm_outputs[[19]]$minimum
#> [1] -1.125
#> 
#> $nlm_outputs[[19]]$estimate
#> [1] 1.249975
#> 
#> $nlm_outputs[[19]]$gradient
#> [1] -8.881965e-12
#> 
#> $nlm_outputs[[19]]$code
#> [1] 1
#> 
#> $nlm_outputs[[19]]$iterations
#> [1] 2
#> 
#> 
#> $nlm_outputs[[20]]
#> $nlm_outputs[[20]]$minimum
#> [1] -1.125
#> 
#> $nlm_outputs[[20]]$estimate
#> [1] -1.249975
#> 
#> $nlm_outputs[[20]]$gradient
#> [1] -1.776393e-12
#> 
#> $nlm_outputs[[20]]$code
#> [1] 1
#> 
#> $nlm_outputs[[20]]$iterations
#> [1] 2
#> 
#> 
#> 
#> $minimize
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "ao"