Alternating Optimization

This function performs alternating optimization of the function f.

Alternating optimization of (high-dimensional) functions is an iterative procedure for optimizing jointly over all parameters by alternately optimizing parameter subsets.

Usage

ao(
  f,
  p,
  ...,
  partition = as.list(1:length(p)),
  base_optimizer = optimizer_optim(),
  iterations = 10,
  tolerance = 1e-06,
  print.level = 0,
  plot = FALSE
)

Arguments

f

A function to be optimized, returning a single numeric. The first argument of f must be a numeric of the length of p followed by any other arguments specified by the ... argument.

p

A numeric vector, the starting parameter values for the optimization.

...

Additional arguments to be passed to f.

partition

A list of vectors of indices of p, specifying the partition of the alternating optimization. The default is as.list(1:length(p)), i.e. each parameter is optimized separately. Parameter indices can be members of multiple subsets.

base_optimizer

An optimizer object, which can be specified via define_optimizer. The default optimizer is optim.

iterations

An integer, the number of iterations through the parameter indices in partitions. The default is 10.

tolerance

A non-negative numeric. The function terminates prematurely if the euclidean distance between the current solution and the one from the last iteration is smaller than tolerance. The default is 1e-6.

print.level

This argument determines the level of printing which is done during the optimization process. Three values (analogue to nlm) can be specified:

0 (the default): no printing

1: initial and final details are printed

2: full tracing information is printed

plot

If TRUE, the parameter updates are plotted. The default is FALSE.

Value

A list containing the following components:

estimate: a numeric, the optimal parameter vector found,

optimum: a numeric, the value of f at estimate,

sequence: a data.frame of the estimates and computation times in the single iterations,

and time, a difftime object, the overall computation time.

See also

Author

Maintainer: Lennart Oelschläger oelschlaeger.lennart@gmail.com (ORCID)

Examples

### alternating optimization separately for x_1 and x_2
### parameter restriction: -5 <= x_1, x_2 <= 5
himmelblau <- function(x) (x[1]^2 + x[2] - 11)^2 + (x[1] + x[2]^2 - 7)^2
ao(
  f = himmelblau, p = c(0,0), partition = list(1, 2), iterations = 10,
  base_optimizer = optimizer_optim(lower = -5, upper = 5, method = "L-BFGS-B")
)
#> $optimum
#> [1] 1.940035e-12
#> 
#> $estimate
#> [1]  3.584428 -1.848126
#> 
#> $sequence
#>    iteration partition         time       p1        p2
#> 1          0         0 0.0000000000 0.000000  0.000000
#> 2          1         1 0.0003039837 3.395691  0.000000
#> 3          1         2 0.0002729893 3.395691 -1.803183
#> 4          2         1 0.0002377033 3.581412 -1.803183
#> 5          2         2 0.0001945496 3.581412 -1.847412
#> 6          3         1 0.0002977848 3.584381 -1.847412
#> 7          3         2 0.0001420975 3.584381 -1.848115
#> 8          4         1 0.0001478195 3.584427 -1.848115
#> 9          4         2 0.0001165867 3.584427 -1.848126
#> 10         5         1 0.0001208782 3.584428 -1.848126
#> 11         5         2 0.0001184940 3.584428 -1.848126
#> 
#> $time
#> Time difference of 0.004497051 secs
#>